Add sortable table util file and handle custom sorting
