+++ app/src/js/components/SortableTable/SortableTable.js
@@ -1,55 +1,25 @@
 import React, {
   useMemo,
   useEffect,
-  forwardRef,
-  useRef,
   useState,
   createRef
 } from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import omit from 'lodash/omit';
-import { defaultOrderByFn, useTable, useResizeColumns, useFlexLayout, useSortBy, useRowSelect, usePagination } from 'react-table';
+import {
+  defaultOrderByFn,
+  useTable,
+  useResizeColumns,
+  useFlexLayout,
+  useSortBy,
+  useRowSelect,
+  usePagination
+} from 'react-table';
 import SimplePagination from '../Pagination/simple-pagination';
 import TableFilters from '../Table/TableFilters';
 import ListFilters from '../ListActions/ListFilters';
-
-const getColumnWidth = (rows, accessor, headerText, originalWidth) => {
-  const maxWidth = 400;
-  const magicSpacing = 10;
-  const cellLength = Math.max(
-    ...rows.map((row) => (`${row.values[accessor]}` || '').length * magicSpacing),
-    headerText.length * magicSpacing,
-    originalWidth,
-  );
-  return Math.min(maxWidth, cellLength);
-};
-
-/**
- * IndeterminateCheckbox
- * @description Component for rendering the header and column checkboxs when canSelect is true
- * Taken from react-table examples
- */
-const IndeterminateCheckbox = forwardRef(
-  ({ indeterminate, title, ...rest }, ref) => {
-    const defaultRef = useRef();
-    const resolvedRef = ref || defaultRef;
-
-    useEffect(() => {
-      resolvedRef.current.indeterminate = indeterminate;
-    }, [resolvedRef, indeterminate]);
-
-    return (
-      <input type="checkbox" ref={resolvedRef} aria-label={title} title={title} {...rest} />
-    );
-  }
-);
-
-IndeterminateCheckbox.propTypes = {
-  indeterminate: PropTypes.any,
-  onChange: PropTypes.func,
-  title: PropTypes.string,
-};
+import { checkInView, getColumnWidth, IndeterminateCheckbox, sortData } from '../../utils/sortable-table';
 
 const SortableTable = ({
   canSelect,
@@ -151,6 +121,7 @@ const SortableTable = ({
   const tableRows = page || rows;
 
   const [sortedRows] = React.useMemo(() => {
+    // we only want to do this if we're already manually sorting but have columns we want to sort client side
     if (!manualSortBy || !sortBy.length) {
       return [tableRows, flatRows];
     }
@@ -161,59 +132,14 @@ const SortableTable = ({
     const availableSortBy = sortBy.filter((sort) => allColumns
       .find((col) => (typeof col.sortMethod === 'function') && (col.id === sort.id)));
 
-    console.log(availableSortBy);
-
+    //  if we don't find any columns with a defined sortMethod, let's bail
     if (!availableSortBy.length) {
       return [tableRows, flatRows];
     }
 
-    const sortData = (_rows) => {
-      // Use the orderByFn to compose multiple sortBy's together.
-      // This will also perform a stable sorting using the row index
-      // if needed.
-      const sortedData = orderByFn(
-        _rows,
-        availableSortBy.map((sort) => {
-          // Support custom sorting methods for each column
-          const column = allColumns.find((d) => d.id === sort.id);
+    const sortedData = sortData({ allColumns, availableSortBy, orderByFn, rows: tableRows, sortedFlatRows });
 
-          if (!column) {
-            throw new Error(
-              `React-Table: Could not find a column with id: ${sort.id} while sorting`
-            );
-          }
-
-          const { sortMethod } = column;
-          // Return the correct sortFn.
-          // This function should always return in ascending order
-          return (a, b) => sortMethod(a, b, sort.id, sort.desc);
-        }),
-        // Map the directions
-        availableSortBy.map((sort) => {
-          // Detect and use the sortInverted option
-          const column = allColumns.find((d) => d.id === sort.id);
-
-          if (column && column.sortInverted) {
-            return sort.desc;
-          }
-
-          return !sort.desc;
-        })
-      );
-
-      // If there are sub-rows, sort them
-      sortedData.forEach((row) => {
-        sortedFlatRows.push(row);
-        if (!row.subRows || row.subRows.length === 0) {
-          return;
-        }
-        row.subRows = sortData(row.subRows);
-      });
-
-      return sortedData;
-    };
-
-    return [sortData(tableRows), sortedFlatRows];
+    return [sortedData, sortedFlatRows];
   }, [manualSortBy, sortBy, tableRows, flatRows, allColumns, orderByFn]);
 
   const includeFilters = typeof getToggleColumnOptions !== 'function';
@@ -359,30 +285,6 @@ const SortableTable = ({
       setRightScrollButtonVisibility({ display: rightScrollButtonVisibility.display, opacity: 0 });
       setRightScrollButtonVisibility({ display: 'none', opacity: 0 });
     }
-  }
-
-  function checkInView(container, element, partial) {
-    if (!container || !element) {
-      return true;
-    }
-
-    // Get container properties
-    const cLeft = container.scrollLeft;
-    const cRight = cLeft + container.clientWidth;
-
-    // Get element properties
-    const eLeft = element.offsetLeft - container.offsetLeft;
-    const eRight = eLeft + element.clientWidth;
-
-    // Check if in view
-    const isTotal = (eLeft >= cLeft && eRight <= cRight);
-    const isPartial = partial && (
-      (eLeft < cLeft && eRight > cLeft) ||
-      (eRight > cRight && eLeft < cRight)
-    );
-
-    // Return outcome
-    return (isTotal || isPartial);
   }
 
   return (
